refactor: move context types to service file
