+++ src/auth/shield/shield.permissions.ts
@@ -17,7 +17,6 @@ import {
   canUpdateGroup,
   hasValidRefreshToken,
   isAuthenticated,
-  isGroupAvatarInPublicFeed,
   isGroupInPublicFeed,
   isGroupMember,
   isOwnPost,
@@ -26,7 +25,6 @@ import {
   isPublicGroup,
   isPublicGroupPost,
   isPublicGroupProposal,
-  isUserAvatarInPublicFeed,
   isUserInPublicFeed,
 } from "./shield.rules";
 
@@ -77,11 +75,7 @@ export const shieldPermissions = shield(
       profilePicture: or(isAuthenticated, isUserInPublicFeed),
     },
     Image: {
-      id: or(
-        isAuthenticated,
-        isUserAvatarInPublicFeed,
-        isGroupAvatarInPublicFeed
-      ),
+      id: or(isUserInPublicFeed, isGroupInPublicFeed),
     },
     Group: {
       id: or(isAuthenticated, isGroupInPublicFeed),
